moved user logc
