+++ src/ws/index.ts
@@ -1,82 +1,31 @@
 import http from 'http';
+import { Action } from 'tic-tac-toe-message';
 import { WebSocketServer } from 'ws';
 
-import { Action } from 'tic-tac-toe-message';
-import { keepAliveTimeout, maxUsers } from '../app.config';
 import { gamesDeleteByUser } from '../game';
-import { User, logUser } from '../user';
+import { userDelete, userGetFromRequest, usersStartHartBeat } from '../user';
+import { createLogger } from '../utils';
 import {
   actionGamesForUsr,
   actionJoinGame,
   actionMove,
   actionNewGame,
-  broadcast,
-  createLogger,
-  getUserFromRequest,
   parseMessage,
 } from './utils';
 
-const users: User[] = [];
-
 const logMsg = createLogger('Message').info;
-
-function userDelete(userId: string) {
-  const idx = users.findIndex(({ id }) => id === userId);
-  if (idx >= 0) {
-    users.splice(idx, 1);
-  }
-}
 
 export function startWsServer(
   server: http.Server<typeof http.IncomingMessage, typeof http.ServerResponse>,
 ) {
   const wsServer = new WebSocketServer({ server });
 
-  setInterval(() => {
-    broadcast(JSON.stringify({ action: Action.KEEP_ALIVE }), users);
-    console.log(
-      'DEBUG:ws:k-a:users',
-      users.map(({ name, id }) => ({ name, id })),
-    );
-    const keepAliveLastTime = Date.now() - keepAliveTimeout;
-    const usersDead = users.filter(({ keepAlive }) => {
-      return keepAlive < keepAliveLastTime;
-    });
-    console.log(
-      'DEBUG:ws:k-a:keepAliveLastTime:deadUsers',
-      keepAliveLastTime,
-      usersDead.map(({ name, id }) => ({ name, id })),
-    );
-    usersDead.forEach(userDead => {
-      userDead.connection.close(4000, 'keep alive timed out');
-      logUser('kick out:', { ...userDead, connection: undefined });
-      gamesDeleteByUser(userDead.id);
-      userDelete(userDead.id);
-    });
-  }, keepAliveTimeout);
+  usersStartHartBeat();
 
+  // ws://localhost:4000?name=John&userId=blah-blah-blah
   wsServer.on('connection', (connection, request) => {
-    // ws://localhost:4000?name=Johnny&userId=blah-blah-blah
-    const { name, userId } = getUserFromRequest(request);
-    const userIdx = users.findIndex(({ id }) => userId === id);
-    let user: User; // TODO: move user state to User model
-    let idx: number = -1; // TODO: move index from state to calc in an util
+    const user = userGetFromRequest(connection, request);
 
-    if (userIdx >= 0) {
-      user = users[userIdx];
-      user.keepAlive = Date.now();
-    } else {
-      user = new User(name as string, connection);
-      if (users.length >= maxUsers) {
-        // TODO: revert with an error instead of connect/close
-        user.connection.close(
-          4001,
-          'number of connections exceeded the limit, please try later.',
-        );
-      } else {
-        idx = users.push(user) - 1;
-      }
-    }
     // log outgoing traffic
     if (process.env.NODE_ENV === 'develop') {
       const send = connection.send;
